+++ src/test/java/org/chucc/vcserver/integration/GraphEventProjectorIT.java
@@ -116,7 +116,7 @@ class GraphEventProjectorIT {
       NewTopic newTopic = new NewTopic(
           topicName,
           kafkaProperties.getPartitions(),
-          (short) kafkaProperties.getReplicationFactor()
+          kafkaProperties.getReplicationFactor()
       );
 
       adminClient.createTopics(Collections.singletonList(newTopic)).all().get();
